+++ src/info.rs
@@ -2,11 +2,11 @@ use super::archive;
 use super::chdman;
 use super::ctrtool;
 use super::database::*;
-use super::flips;
 use super::nsz;
 use super::progress::*;
 use super::rvz;
 use super::xdelta;
+use super::xps;
 use super::xso;
 use anyhow::Result;
 use clap::Command;
@@ -27,7 +27,7 @@ pub async fn main(connection: &mut SqliteConnection, progress_bar: &ProgressBar)
         ("sevenz-rust2", archive::get_version().await),
         ("chdman", chdman::get_version().await),
         ("ctrtool", ctrtool::get_version().await),
-        ("flips", flips::get_version().await),
+        ("xps-rs", xps::get_version().await),
         ("xso-rs", xso::get_version().await),
         ("nsz-rs", nsz::get_version().await),
         ("xdelta-rs", xdelta::get_version().await),